Fixed possible loss of data warning when compiling for x64.
When compiling for x86 this is fine because Py_ssize_t is typedef'd to ``int``, however on x64 it is typedef'd to ``long long`` and so ``int`` here is a downcast and can result in "data loss". As such it is best to use ``Py_ssize_t`` instead here on these and permanently silent these warnings as the type would of these variables would automatically adjust based on the definition of ``Py_ssize_t``.